+++ scripts/ci-test-deps.sh
@@ -6,8 +6,8 @@ BITCOIN_FILENAME=bitcoin-$BITCOIN_VERSION-x86_64-linux-gnu.tar.gz
 BITCOIN_URL=https://bitcoincore.org/bin/bitcoin-core-$BITCOIN_VERSION/$BITCOIN_FILENAME
 BITCOIN_SHA256=35ec10f87b6bc1e44fd9cd1157e5dfa483eaf14d7d9a9c274774539e7824c427
 
-ELECTRUM_URL=https://download.electrum.org/3.3.8/electrum-3.3.8-x86_64.AppImage
-ELECTRUM_SHA256=fd2526f7a37530bb655375d3c95ce6a51e3c2beb83ff88a54e28aa365033794a
+ELECTRUM_URL=https://download.electrum.org/4.0.3/electrum-4.0.3-x86_64.AppImage
+ELECTRUM_SHA256=512b58c437847048a9629cb6bf2eb786b8969a1e17b7b51510b11672c9b29fc7
 
 mkdir -p /opt/bin /opt/bitcoin